changed flaot to double

+++ Trivia/IDatabase.h
@@ -16,7 +16,7 @@ public:
 	virtual bool doesPasswordMatch(const std::string& username, const std::string& password) = 0;
 
 	/* Statistics Queries */
-	virtual float getPlayerAverageAnswerTime(const std::string& userName) = 0;
+	virtual double getPlayerAverageAnswerTime(const std::string& userName) = 0;
 	virtual int getNumOfCorrectAnswers(const std::string& userName) = 0;
 	virtual int getNumOfTotalAnswers(const std::string& userName) = 0;
 	virtual int getNumOfPlayerGames(const std::string& userName) = 0;
